requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
playwright>=1.46.0
openai>=1.14.0
python-dotenv>=1.0.0
pydantic>=2.7.0
aiofiles>=23.2.1
